pyproject.toml
[project]
name = "socialdigest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psutil>=7.0.0",
    "python-dotenv>=1.0.1",
    "selenium>=4.28.1",
    "setuptools>=75.8.0",
    "tweepy>=4.15.0",
    "twikit>=2.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-mock>=3.14.0",
]
